Normal벡터 normalize 오류 수정 + Q2 진행

+++ EmptyViewer/Surface.h
@@ -26,5 +26,5 @@ protected:
 	vec3 ka, kd, ks;
 
 	float Ia = 1;
-	float specularPower;
+	int specularPower;
 };
